+++ test/Store_test.dart
@@ -4,7 +4,6 @@
 
 import 'dart:async';
 import 'dart:utf';
-import 'dart:typed_data';
 import 'package:unittest/unittest.dart';
 import 'package:memcached_client/memcached_client.dart';
 import 'MemcachedTestUtil.dart' as m;
@@ -40,7 +39,7 @@ void testCas(MemcachedClient client) {
   f1.then((v) {
     expect(decodeUtf8(v.data), equals('中文'));
     expect(v.cas, isNotNull);
-    expect(client.set('key0', encodeUtf8('val0'), v.cas), completion(isTrue));
+    expect(client.set('key0', encodeUtf8('val0'), cas:v.cas), completion(isTrue));
   });
   expect(f1, completes);
 }